refactor(oracle): migrate rip to per-protocol plugins
