+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -50,7 +50,6 @@ KNOWN_UNMIGRATED: frozenset[str] = frozenset(
         "esp",
         "ikev2",
         "radiotap",
-        "rip",
         "ripng",
         "rsn",
     }
